pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airarmy"
version = "0.1.0"
description = "Your personal AI company — a multi-agent system with security-first design"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.scripts]
airarmy = "airarmy.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["airarmy*"]
